% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_plots.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot abundance profiles for MAGs at different buoyant densities}
\usage{
plot_abundance(incorporator_list, fractions, mag_tab)
}
\arguments{
\item{incorporator_list}{A tibble consisting of two columns
\itemize{
\item First column consists of incorporator feature ID.
\item Second column consists of the incorporator taxonomy
}}

\item{fractions}{Fractions metadata file consisting of
\itemize{
\item Replicate: Depends on how many replicates the study has
\item Fractions: Typically in the range of 2-24
\item Buoyant_density: As calculated from the refractometer for each fraction and replicate
\item Isotope: "12C", "13C", "14N", "15N" etc.
\item DNA_concentration
\item Sample: In the format "'isotope'\emph{rep}#\emph{fraction}#".
For instance, "12C_rep_1_fraction_1"
}}

\item{mag_tab}{A tibble of feature absolute abundances
Each column consists of the scaled abundances of the features. Row names idenitfy the Feature}
}
\value{
image files with the following formats
\itemize{
\item "mean_abundance_" prefix is for the plots visualizing the abundance estimates of features
over the buoyant density gradient
\item "Rep_1_abundance" prefix is for the plots visualizing the abundance estimates of features
from replicate 1 over the buoyant density gradient
\item "Rep_2_abundance" prefix is for the plots visualizing the abundance estimates of features
from replicate 2 over the buoyant density gradient
\item "Rep_3_abundance" prefix is for the plots visualizing the abundance estimates of features from replicate 3 over the buoyant density gradient
}
}
\description{
This function provides the plots for abundance vs Buoyant density of incorporators. The plots are saved in the directory "abundance_plots" in the current working directory of the "qSIP_analysis.rmd" pipeline. The goal of plotting individual replicates separately is to identify any anomalies, if any, regarding the changes in the scaled abundances across the buoyant density gradient which may not be obvious in the plots with the mean and standard deviation of abundances. These plots could help in debugging
}
